refactor(web): apiRequest, token client-server 분리
