+++ apps/web/src/api/apiRequest.client.ts
@@ -1,0 +1,8 @@
+import { getRefreshToken } from '@/auth/apis/refresh';
+import { getAccessToken } from '@/auth/token.client';
+import { createApiRequest } from '@/api/apiRequest.base';
+
+export const apiRequest = createApiRequest({
+  getAccessToken,
+  getRefreshToken,
+});
