+++ apps/web/src/api/apiRequest.server.ts
@@ -1,0 +1,10 @@
+import 'server-only';
+
+import { getRefreshToken } from '@/auth/apis/refresh';
+import { getAccessToken } from '@/auth/token.server';
+import { createApiRequest } from '@/api/apiRequest.base';
+
+export const apiRequest = createApiRequest({
+  getAccessToken,
+  getRefreshToken,
+});
